Adding the classes for the multi-peer connectivity: properties MCSession, MCPeerID, MCAdvertiserAssistant, and MCBrowserViewController class.
Implementing the methods didReceive stream, didStartReceiving, didFinishReceiving to make the app be compatible with the above protocols (the methods will be empty as they are required but not useful in our case). Also implemented browserViewControllerDidFinish and browserViewControllerWasCancelled.

Configured the didChange method in order to visualise in the console the status of the connection.